refactor: use kanvas activity
