--- graphql/schemas/Intelligence/followUp.graphql
@@ -1,45 +1,0 @@
-type FollowUpPromptTemplate {
-    id: ID!
-    name: String
-    content: String
-}
-
-type FollowUpPromptSession {
-    session_id: ID!
-    session_uuid: String!
-    channel: String!
-    template: FollowUpPromptTemplate
-    prompt: String
-    day: Float
-}
-
-type FollowUpInfo {
-    id: ID!
-    name: String!
-    config: Mixed
-}
-
-type FollowUpDayInfo {
-    id: ID!
-    name: String
-    time_value: Int
-    time_unit: String
-    send_message: Boolean
-}
-
-type FollowUpPromptResponse {
-    lead_id: ID!
-    stage_id: ID
-    stage_name: String
-    is_simulation: Boolean!
-    follow_up: FollowUpInfo
-    follow_up_day: FollowUpDayInfo
-    sessions: [FollowUpPromptSession!]!
-}
-
-extend type Query @guard {
-    agentFollowUpPrompts(lead_id: ID!, stage_id: ID): FollowUpPromptResponse!
-        @field(
-            resolver: "App\\GraphQL\\Intelligence\\Queries\\FollowUpPromptQuery@getAgentFollowUpPrompts"
-        )
-}
