+++ src/Domains/Workflow/Enums/WorkflowEnum.php
@@ -46,6 +46,7 @@ enum WorkflowEnum: string
     case TRIGGER_AI = 'trigger-ai';
     case AFTER_LEGACY_FEED = 'after-legacy-feed';
     case SUBSCRIPTION_LIFECYCLE = 'subscription-lifecycle';
+    case FOLLOW_UP_PROMPT = 'follow-up-prompt';
 
     /**
      * Get the enum case by its value.
